+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2.0.3
+* Remove `WebSocketController.plugin`.
+
 # 2.0.2
 * Update `stream_channel` to `2.0.0`.
 * Use `angel_framework^@2.0.0-rc.0`.